test(litellm): split retry patch coverage
